+++ backend/src/stages/mod.rs
@@ -1,5 +1,7 @@
 pub mod exporting;
 pub mod fetching;
+pub mod fetching_coop;
+pub mod fetching_sp;
 pub mod uploading;
 pub mod uploading_coop;
 pub mod uploading_sp;
